Improve code quality and test maintainability
- Extract MockOAuthToken class to eliminate duplicate anonymous class implementations
- Make all test classes final following modern PHP best practices
- Fix PHPStan generic-baseline.neon path configuration for CI pipeline
