+++ tests/FunctionalTest.php
@@ -9,7 +9,7 @@ use Symfony\Component\Config\Loader\LoaderInterface;
 use Symfony\Component\DependencyInjection\ContainerBuilder;
 use Symfony\Component\HttpKernel\Kernel;
 
-class FunctionalTest extends TestCase
+final class FunctionalTest extends TestCase
 {
     public function testServiceWiring(): void
     {
