+++ tests/OAuthSubscriberFactoryTest.php
@@ -7,7 +7,7 @@ use Calliostro\DiscogsBundle\OAuthTokenProviderInterface;
 use GuzzleHttp\Subscriber\Oauth\Oauth1;
 use PHPUnit\Framework\TestCase;
 
-class OAuthSubscriberFactoryTest extends TestCase
+final class OAuthSubscriberFactoryTest extends TestCase
 {
     public function testFactory(): void
     {
